// src/lib/build-your-own-x-data.generated.ts
export const buildYourOwnXGeneratedCategories = [
  {
    "slug": "distributed-systems",
    "name": "Distributed Systems",
    "count": 1
  },
  {
    "slug": "3d-renderer",
    "name": "3D Renderer",
    "count": 11
  },
  {
    "slug": "ai-model",
    "name": "AI Model",
    "count": 3
  },
  {
    "slug": "augmented-reality",
    "name": "Augmented Reality",
    "count": 6
  },
  {
    "slug": "bittorrent-client",
    "name": "BitTorrent Client",
    "count": 5
  },
  {
    "slug": "blockchain-cryptocurrency",
    "name": "Blockchain / Cryptocurrency",
    "count": 21
  },
  {
    "slug": "bot",
    "name": "Bot",
    "count": 15
  },
  {
    "slug": "command-line-tool",
    "name": "Command-Line Tool",
    "count": 9
  },
  {
    "slug": "database",
    "name": "Database",
    "count": 13
  },
  {
    "slug": "docker",
    "name": "Docker",
    "count": 6
  },
  {
    "slug": "emulator-virtual-machine",
    "name": "Emulator / Virtual Machine",
    "count": 13
  },
  {
    "slug": "front-end-framework-library",
    "name": "Front-end Framework / Library",
    "count": 14
  },
  {
    "slug": "game",
    "name": "Game",
    "count": 34
  },
  {
    "slug": "git",
    "name": "Git",
    "count": 7
  },
  {
    "slug": "memory-allocator",
    "name": "Memory Allocator",
    "count": 1
  },
  {
    "slug": "network-stack",
    "name": "Network Stack",
    "count": 4
  },
  {
    "slug": "neural-network",
    "name": "Neural Network",
    "count": 17
  },
  {
    "slug": "operating-system",
    "name": "Operating System",
    "count": 19
  },
  {
    "slug": "physics-engine",
    "name": "Physics Engine",
    "count": 7
  },
  {
    "slug": "processor",
    "name": "Processor",
    "count": 1
  },
  {
    "slug": "programming-language",
    "name": "Programming Language",
    "count": 41
  },
  {
    "slug": "regex-engine",
    "name": "Regex Engine",
    "count": 9
  },
  {
    "slug": "search-engine",
    "name": "Search Engine",
    "count": 6
  },
  {
    "slug": "shell",
    "name": "Shell",
    "count": 6
  },
  {
    "slug": "template-engine",
    "name": "Template Engine",
    "count": 5
  },
  {
    "slug": "text-editor",
    "name": "Text Editor",
    "count": 5
  },
  {
    "slug": "visual-recognition-system",
    "name": "Visual Recognition System",
    "count": 2
  },
  {
    "slug": "voxel-engine",
    "name": "Voxel Engine",
    "count": 1
  },
  {
    "slug": "web-browser",
    "name": "Web Browser",
    "count": 2
  },
  {
    "slug": "web-server",
    "name": "Web Server",
    "count": 71
  }
] as const;